[feat]: fix typo

+++ app/src/main/java/com/igniter/ffmpegtest/data/data_source/FFmpegSolution.kt
@@ -17,7 +17,7 @@ object FFmpegSolution {
      */
     external fun capture(
         videoPath: String,
-        startTimeInS: Int,
+        startTimeInS: Long,
         startPos: Int,
         totalNum: Int,
         enableMultiThread: Boolean,
